Reorganize tests
Moved common stuff under "common" package.
